// client/src/app/services/swagger-expresscart-client/model/createProductRequest.ts
/**
 * Expresscart
 * Expresscart API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateProductRequest {
    title: string;
    description: string;
    price: number;
    file?: any | null;
}